Dev sachin

+++ gempyp/libs/common.py
@@ -147,15 +147,19 @@ def moduleImports(file_name):
             return e
 
 def download_common_file(file_name,headers=None):
-    try:
+     try:
         if file_name and (file_name.__contains__('S3')):
             logging.info("File is from S3")
-            fileContent=download_from_s3(api=file_name.replace("S3:",""),username=headers.get("username",None),bridge_token=headers.get("bridge_token",None))
-            file_name = os.path.join(file_name.split(":")[-1])
-            with open(file_name, "w+") as fp:
-                fp.seek(0)
-                fp.write(fileContent)
-                fp.truncate()
+            response=download_from_s3(api=file_name.replace("S3:",""),username=headers.get("username",None),bridge_token=headers.get("bridge_token",None))
+            if(response.status_code==200):
+                file_name = os.path.join(file_name.split(":")[-1])
+                with open(file_name, "w+") as fp:
+                    fp.seek(0)
+                    fp.write(response.text)
+                    fp.truncate()
+            else:
+                logging.info(response.status_code)
+                logging.info(response.text)
         elif file_name and (file_name.__contains__('GIT')):
             logging.info("File is from GIT")
             list_url=file_name.split(":")
@@ -164,7 +168,7 @@ def download_common_file(file_name,headers=None):
             else:
                 file_name=fetchFileFromGit(list_url[2],list_url[3])
         return file_name
-    except Exception as e:
+     except Exception as e:
         traceback.print_exc()
         return e
